Testnet

See merge request lidonation/lidonation!422

+++ src/www.lidonation.com/var/www/app/Http/Controllers/ProjectCatalyst/CatalystMyBookmarksController.php
@@ -13,6 +13,7 @@ use App\Services\ExportModelService;
 use App\Exports\BookmarksCollectionExport;
 use App\Http\Resources\BookmarkCollectionResource;
 use Google\Service\ShoppingContent\Resource\Collections;
+use Illuminate\Support\Facades\DB;
 
 class CatalystMyBookmarksController extends Controller
 {
@@ -20,6 +21,8 @@ class CatalystMyBookmarksController extends Controller
 
     public function createItem(Request $request)
     {
+        DB::beginTransaction();
+        try{
         $modelTable = $request->get('model_type');
         $data = new Fluent($request->validate([
             'model_id' => "required|exists:{$modelTable},id",
@@ -56,12 +59,18 @@ class CatalystMyBookmarksController extends Controller
         $item->content = $data->content;
         $item->link = $data->link;
         $item->save();
+         
+        DB::commit();
 
         $collection->refresh();
         $collection->load(['items']);
 
         return new BookmarkCollectionResource($collection);
-    }
+    }catch(\Exception $e) {
+        DB::rollback();
+        throw new \Exception("There was an Error");
+    }    
+}
 
     public function createCollection()
     {
@@ -92,15 +101,23 @@ class CatalystMyBookmarksController extends Controller
     }
 
     function deleteCollection(Request $request)
-    {
+    {   
         $collection = BookmarkCollection::byHash($request->hash);
+        $this->authorize('delete', $collection);
+
         $collection->items()->delete();
         $collection->delete();
     }
 
+    function deleteItem( BookmarkItem $bookmarkItem)
+    {   
+        $this->authorize('delete', $bookmarkItem);
+        $bookmarkItem->delete();
+    }
+
     public function exportBookmarks(Request $request) 
     {
-        
+
         $collection = BookmarkCollection::byHash($request->hash);
         $itemsArr=$collection->items()->pluck('id');
